refactor so there is a calculation
