+++ BE/test.ts
@@ -1,3 +1,3 @@
-import { parseInputMovement } from "./lib/penguin/movementHandler";
+import { parseInputMovement } from "./lib/penguin/movementHandlers";
 
 parseInputMovement({ penguinId: 'a37deb36-58a2-4cea-8a4b-e2ebc024f1b5', clickDestPos: [9, 6], arrowKeyPressed: null })
